requirements.txt
pandas==2.2.2
openpyxl==3.1.2
langchain==0.2.16
langchain-community==0.2.16
langchain-ollama==0.1.3
chromadb==0.5.5
reportlab==4.2.2
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6